Dial Controller and Player Interaction v1
In this version, I have slighlt changed the dial controller already to v2 but nothing else yet. the main problem with v1 is that player interaction wasn't usable using keyboard and mouse and toggling the dial with "E" did not work. Axis was also not aligned for the stirrer as the stirrer rotated around the jart tester machine