Ethan update - able to turn stirrer in place and change RPM using textbox
was able to get the stirrers to turn in place and replicated 6 times over for the jar tester. in addition, he is working on getting the water to change color based on RPM, but will need to see how this works in accordance with the lab that is being modelled.